+++ innerjoinus-article-reply-source/src/main/java/org/thisdote/innerjoinus/articlereply/reportedarticle/command/controller/CommandReportedArticleController.java
@@ -1,4 +1,5 @@
 package org.thisdote.innerjoinus.articlereply.reportedarticle.command.controller;
 
 public class CommandReportedArticleController {
+
 }
