+++ innerjoinus-article-reply-source/src/main/java/org/thisdote/innerjoinus/articlereply/reportedarticle/dto/ReportedArticle.java
@@ -1,0 +1,16 @@
+package org.thisdote.innerjoinus.articlereply.reportedarticle.dto;
+
+import lombok.Data;
+
+import java.util.Date;
+
+@Data
+public class ReportedArticle {
+    private int reportArticleId;
+    private int articleId;
+    private int userCode;
+    private Date reportDate;
+    private String reportContent;
+    private int reportReportedCount;
+    private int reportType;
+}
