+++ innerjoinus-article-reply-source/src/main/java/org/thisdote/innerjoinus/articlereply/reportedreply/command/controller/CommandReportedReplyController.java
@@ -1,0 +1,5 @@
+package org.thisdote.innerjoinus.articlereply.reportedreply.command.controller;
+
+public class CommandReportedReplyController {
+    
+}
