+++ innerjoinus-article-reply-source/src/main/java/org/thisdote/innerjoinus/articlereply/reportedreply/command/repository/CommandReportedReplyRepository.java
@@ -1,0 +1,4 @@
+package org.thisdote.innerjoinus.articlereply.reportedreply.command.repository;
+
+public class CommandReportedReplyRepository {
+}
