+++ innerjoinus-article-reply-source/src/main/java/org/thisdote/innerjoinus/articlereply/reportedreply/command/service/CommandReportedReplyServiceImpl.java
@@ -1,0 +1,4 @@
+package org.thisdote.innerjoinus.articlereply.reportedreply.command.service;
+
+public class CommandReportedReplyServiceImpl {
+}
