+++ innerjoinus-article-reply-source/src/main/java/org/thisdote/innerjoinus/articlereply/reportedreply/command/vo/Response.java
@@ -1,0 +1,4 @@
+package org.thisdote.innerjoinus.articlereply.reportedreply.command.vo;
+
+public class Response {
+}
