--- innerjoinus-article-reply-source/src/main/java/org/thisdote/innerjoinus/articlereply/reportedreply/controller/CommandReportedReplyController.java
@@ -1,4 +1,0 @@
-package org.thisdote.innerjoinus.articlereply.reportedreply.controller;
-
-public class CommandReportedReplyController {
-}
