+++ innerjoinus-article-reply-source/src/main/java/org/thisdote/innerjoinus/articlereply/reportedreply/dto/ArticleReplyDTO.java
@@ -1,0 +1,16 @@
+package org.thisdote.innerjoinus.articlereply.reportedreply.dto;
+
+import lombok.Data;
+
+import java.util.Date;
+
+@Data
+public class ArticleReplyDTO {
+    private int reportReplyId;
+    private int articleId;
+    private int userCode;
+    private Date reportDate;
+    private String reportContent;
+    private int reportReportedCount;
+    private int reportType;
+}
