--- innerjoinus-article-reply-source/src/main/java/org/thisdote/innerjoinus/articlereply/reportedreply/repository/CommandReportedReplyRepository.java
@@ -1,4 +1,0 @@
-package org.thisdote.innerjoinus.articlereply.reportedreply.repository;
-
-public class CommandReportedReplyRepository {
-}
