--- innerjoinus-article-reply-source/src/main/java/org/thisdote/innerjoinus/articlereply/reportedreply/service/CommandReportedReplyService.java
@@ -1,4 +1,0 @@
-package org.thisdote.innerjoinus.articlereply.reportedreply.service;
-
-public interface CommandReportedReplyService {
-}
