--- innerjoinus-article-reply-source/src/main/java/org/thisdote/innerjoinus/articlereply/reportedreply/service/CommandReportedReplyServiceImpl.java
@@ -1,4 +1,0 @@
-package org.thisdote.innerjoinus.articlereply.reportedreply.service;
-
-public class CommandReportedReplyServiceImpl {
-}
